Check if input alias name is new before inserting
New alias with the same name as an existing alias will not be
inserted into the hashtable. Earlier, the user was allowed to
create new alias with the same name as an existing one and this
resulted in seg faults. This commit will print a error message
when the user tries to create a new alias with an existing name.